rename use user modals composable
